Refactoring: - Removing case class RegexDictionary
