+++ src/main/scala/com/johnsnowlabs/nlp/annotators/DeIdentificationModel.scala
@@ -1,8 +1,8 @@
-package com.johnsnowlabs.nlp.annotators.anonymizer
+package com.johnsnowlabs.nlp.annotators
 
 import com.johnsnowlabs.nlp.AnnotatorType.{CHUNK, DOCUMENT, TOKEN}
 import com.johnsnowlabs.nlp.annotators.common.{IndexedToken, SentenceSplit, TokenizedWithSentence}
-import com.johnsnowlabs.nlp.serialization.StructFeature
+import com.johnsnowlabs.nlp.serialization.MapFeature
 import com.johnsnowlabs.nlp.{Annotation, AnnotatorModel}
 import org.apache.spark.ml.util.{DefaultParamsReadable, Identifiable}
 
@@ -15,17 +15,11 @@ class DeIdentificationModel(override val uid: String) extends AnnotatorModel[DeI
   override val annotatorType: AnnotatorType = DOCUMENT
   override val requiredAnnotatorTypes: Array[AnnotatorType] = Array(DOCUMENT, TOKEN, CHUNK)
 
-  // val regexPatternsDictionary: MapFeature[String, Array[String]] = new MapFeature(this, "regexPatternsDictionary")
+  val regexPatternsDictionary: MapFeature[String, Array[String]] = new MapFeature(this, "regexPatternsDictionary")
 
-  val regexPatternsDictionary: StructFeature[RegexPatternsDictionary] =
-    new StructFeature[RegexPatternsDictionary](this, "regexPatternsDictionary")
+  def setRegexPatternsDictionary(value: Map[String, Array[String]]): this.type = set(regexPatternsDictionary, value)
 
-  //def setRegexPatternsDictionary(value: Map[String, Array[String]]): this.type = set(regexPatternsDictionary, value)
-
-  def setRegexPatternsDictionary(value: RegexPatternsDictionary): this.type = set(regexPatternsDictionary, value)
-
-  //def getRegexPatternsDictionary: Map[String, Array[String]] = $$(regexPatternsDictionary)
-  def getRegexPatternsDictionary: RegexPatternsDictionary = $$(regexPatternsDictionary)
+  def getRegexPatternsDictionary: Map[String, Array[String]] = $$(regexPatternsDictionary)
 
   private lazy val regexDictionary = getRegexPatternsDictionary
 
@@ -47,12 +41,11 @@ class DeIdentificationModel(override val uid: String) extends AnnotatorModel[DeI
 
     var regexEntities = new ListBuffer[Annotation]()
 
-    //if (regexDictionary.isEmpty){
-    if (regexDictionary.dictionary.isEmpty){
+    if (regexDictionary.isEmpty){
       return Seq()
     }
 
-    for ((entity, regexPatterns) <- regexDictionary.dictionary){
+    for ((entity, regexPatterns) <- regexDictionary){
       tokensSentence.foreach{tokenSentence =>
          if (isRegexMatch(tokenSentence.token, regexPatterns)){
