--- src/main/scala/com/johnsnowlabs/nlp/annotators/anonymizer/RegexPatternsDictionary.scala
@@ -1,3 +1,0 @@
-package com.johnsnowlabs.nlp.annotators.anonymizer
-
-case class RegexPatternsDictionary(dictionary: Map[String, Array[String]]) extends Serializable{}
